backend/requirements.txt
fastapi
uvicorn
qdrant-client
transformers
torch
cohere
google-genai